requirements.txt
asgiref==3.2.3
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
coreapi==2.3.3
coreapi-cli==1.0.9
coreschema==0.0.4
Django==3.0.3
django-cors-headers==3.2.1
django-rest-swagger==2.2.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
get==2019.4.13
idna==2.9
itypes==1.1.0
Jinja2==2.11.1
lxml==4.5.0
MarkupSafe==1.1.1
mysqlclient==1.4.6
openapi-codec==1.3.2
Pillow==7.0.0
post==2019.4.13
public==2019.4.13
PyJWT==1.7.1
pytz==2019.3
query-string==2019.4.13
requests==2.23.0
simplejson==3.17.0
sqlparse==0.3.0
uritemplate==3.0.1
urllib3==1.25.8
